split compactblock and sct into separate modules
