+++ component/src/action_handler/actions/swap_claim.rs
@@ -7,8 +7,7 @@ use penumbra_storage::{StateRead, StateWrite};
 use penumbra_transaction::{action::SwapClaim, Transaction};
 
 use crate::{
-    action_handler::ActionHandler,
-    shielded_pool::{NoteManager, StateReadExt as _},
+    action_handler::ActionHandler, sct::view::StateReadExt as _, shielded_pool::NoteManager,
     stubdex::StateReadExt as _,
 };
