+++ component/src/action_handler/transaction/stateful.rs
@@ -3,7 +3,8 @@ use penumbra_chain::StateReadExt as _;
 use penumbra_storage::StateRead;
 use penumbra_transaction::Transaction;
 
-use crate::shielded_pool::{consensus_rules, StateReadExt as _};
+use crate::sct::view::StateReadExt as _;
+use crate::shielded_pool::consensus_rules;
 
 pub(super) async fn claimed_anchor_is_valid<S: StateRead>(
     state: S,
