+++ component/src/lib.rs
@@ -15,10 +15,12 @@ pub use mock_client::MockClient;
 pub use temp_storage_ext::TempStorageExt;
 
 pub mod app;
+pub mod compactblock;
 pub mod dao;
 pub mod dex;
 pub mod governance;
 pub mod ibc;
+pub mod sct;
 pub mod shielded_pool;
 pub mod stake;
 pub mod stubdex;
