+++ component/src/shielded_pool/mod.rs
@@ -8,7 +8,6 @@ mod supply;
 pub mod state_key;
 
 pub use self::metrics::register_metrics;
-pub(crate) use component::StateWriteExt;
-pub use component::{ShieldedPool, StateReadExt};
+pub use component::ShieldedPool;
 pub use note_manager::NoteManager;
 pub use supply::{SupplyRead, SupplyWrite};
